LayoutTests/media/controls/showControlsButton.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../media-file.js"></script>
        <script src="../video-test.js"></script>
        <script src="controls-test-helpers.js"></script>
        <script>
            var video;
            var currentStatus;
            var showControlsButtonInfo;

            function init()
            {
                findMediaElement();
                video.src = findMediaFile("video", "../content/test");

                waitForEvent("canplaythrough", canplaythrough);
                waitForEventAndFail("error");
            }

            function canplaythrough()
            {
                if (!window.testRunner)
                    return;

                currentStatus = JSON.parse(internals.getCurrentMediaControlsStatusForElement(video));
                if (currentStatus) {
                    if (currentStatus.idiom == "apple") {
                        showControlsButtonInfo = statusForControlsElement(currentStatus, "Show Controls");
                        if (showControlsButtonInfo) {
                            testExpected("showControlsButtonInfo.name", "Show Controls");
                            testExpected("showControlsButtonInfo.hidden", true);
                        }
                    }
                }
                endTest();
            }
        </script>
    </head>
    <body onload="init()">
        <p>This tests for the existence of a hidden Show Controls button.</p>
        <p>This test only runs in DRT!</p>
        <video controls></video>
    </body>
</html>
