LayoutTests/media/controls/showControlsButton-expected.txt

This tests for the existence of a hidden Show Controls button.

This test only runs in DRT!


EVENT(canplaythrough)
EXPECTED (showControlsButtonInfo.name == 'Show Controls') OK
EXPECTED (showControlsButtonInfo.hidden == 'true') OK
END OF TEST
